refactor(evaluations): separate shared runtime from use cases
Move the reusable execution engine under runtime/ and group feature entrypoints under use-cases/ so alignment, live, and optimization flows share the same evaluation surface.
