+++ apps/workflows/src/activities/execute-evaluation-script.test.ts
@@ -1,10 +1,10 @@
 import {
-  CONVERSATION_PLACEHOLDER,
-  extractPromptFromScript,
-  formatConversationForPrompt,
+  EVALUATION_CONVERSATION_PLACEHOLDER,
+  extractPromptFromEvaluationScript,
+  formatEvaluationConversationForPrompt,
   generateBaselinePromptText,
   validateEvaluationScript,
-  wrapPromptAsScript,
+  wrapPromptAsEvaluationScript,
 } from "@domain/evaluations"
 import { describe, expect, it } from "vitest"
 
@@ -14,45 +14,45 @@ const DUMMY_CONVERSATION = [
 ]
 
 describe("MVP script template helpers", () => {
-  describe("wrapPromptAsScript / extractPromptFromScript", () => {
+  describe("wrapPromptAsEvaluationScript / extractPromptFromEvaluationScript", () => {
     it("round-trips a simple prompt", () => {
       const prompt = "Evaluate the conversation for issues."
-      const script = wrapPromptAsScript(prompt)
-      const extracted = extractPromptFromScript(script)
+      const script = wrapPromptAsEvaluationScript(prompt)
+      const extracted = extractPromptFromEvaluationScript(script)
       expect(extracted).toBe(prompt)
     })
 
     it("round-trips a prompt with conversation placeholder", () => {
-      const prompt = `Analyze this conversation:\n${CONVERSATION_PLACEHOLDER}\n\nDoes it contain issues?`
-      const script = wrapPromptAsScript(prompt)
-      const extracted = extractPromptFromScript(script)
+      const prompt = `Analyze this conversation:\n${EVALUATION_CONVERSATION_PLACEHOLDER}\n\nDoes it contain issues?`
+      const script = wrapPromptAsEvaluationScript(prompt)
+      const extracted = extractPromptFromEvaluationScript(script)
       expect(extracted).toBe(prompt)
     })
 
     it("returns null for a script that does not match the template", () => {
-      const result = extractPromptFromScript("return Passed(1, 'looks good')")
+      const result = extractPromptFromEvaluationScript("return Passed(1, 'looks good')")
       expect(result).toBeNull()
     })
 
     it("returns null for an empty string", () => {
-      expect(extractPromptFromScript("")).toBeNull()
+      expect(extractPromptFromEvaluationScript("")).toBeNull()
     })
 
     it("returns null for a partial template match", () => {
-      const script = wrapPromptAsScript("hello")
+      const script = wrapPromptAsEvaluationScript("hello")
       const mangled = script.slice(0, script.length - 5)
-      expect(extractPromptFromScript(mangled)).toBeNull()
+      expect(extractPromptFromEvaluationScript(mangled)).toBeNull()
     })
   })
 
   describe("validateEvaluationScript", () => {
     it("accepts a valid template with conversation placeholder", () => {
-      const script = wrapPromptAsScript(`Check this: ${CONVERSATION_PLACEHOLDER}`)
+      const script = wrapPromptAsEvaluationScript(`Check this: ${EVALUATION_CONVERSATION_PLACEHOLDER}`)
       expect(validateEvaluationScript(script)).toBe(true)
     })
 
     it("accepts a valid template without any interpolation (static prompt)", () => {
-      const script = wrapPromptAsScript("Is the response helpful?")
+      const script = wrapPromptAsEvaluationScript("Is the response helpful?")
       expect(validateEvaluationScript(script)).toBe(true)
     })
 
@@ -62,34 +62,36 @@ describe("MVP script template helpers", () => {
 
     it("rejects a prompt containing forbidden interpolations", () => {
       const forbidden = ["Issue: ${", "issue.name}"].join("")
-      const script = wrapPromptAsScript(forbidden)
+      const script = wrapPromptAsEvaluationScript(forbidden)
       expect(validateEvaluationScript(script)).toBe(false)
     })
 
     it("rejects a prompt containing backticks", () => {
-      const script = wrapPromptAsScript("Use `code` here")
+      const script = wrapPromptAsEvaluationScript("Use `code` here")
       expect(validateEvaluationScript(script)).toBe(false)
     })
 
     it("rejects a prompt with multiple different interpolations", () => {
       const otherPlaceholder = ["${", "other}"].join("")
-      const script = wrapPromptAsScript(`${CONVERSATION_PLACEHOLDER} and ${otherPlaceholder}`)
+      const script = wrapPromptAsEvaluationScript(
+        `${EVALUATION_CONVERSATION_PLACEHOLDER} and ${otherPlaceholder}`,
+      )
       expect(validateEvaluationScript(script)).toBe(false)
     })
   })
 
-  describe("formatConversationForPrompt", () => {
+  describe("formatEvaluationConversationForPrompt", () => {
     it("formats conversation messages into readable text", () => {
-      const result = formatConversationForPrompt(DUMMY_CONVERSATION)
+      const result = formatEvaluationConversationForPrompt(DUMMY_CONVERSATION)
       expect(result).toBe("[user] What is my API key?\n[assistant] Your API key is sk-live-123.")
     })
 
     it("handles an empty conversation", () => {
-      expect(formatConversationForPrompt([])).toBe("")
+      expect(formatEvaluationConversationForPrompt([])).toBe("")
     })
 
     it("handles a single message", () => {
-      const result = formatConversationForPrompt([{ role: "system", content: "You are helpful." }])
+      const result = formatEvaluationConversationForPrompt([{ role: "system", content: "You are helpful." }])
       expect(result).toBe("[system] You are helpful.")
     })
   })
@@ -103,12 +105,12 @@ describe("MVP script template helpers", () => {
 
     it("includes the conversation placeholder", () => {
       const prompt = generateBaselinePromptText("Test Issue", "Test description")
-      expect(prompt).toContain(CONVERSATION_PLACEHOLDER)
+      expect(prompt).toContain(EVALUATION_CONVERSATION_PLACEHOLDER)
     })
 
     it("produces a valid script when wrapped", () => {
       const prompt = generateBaselinePromptText("Test Issue", "Test description")
-      const script = wrapPromptAsScript(prompt)
+      const script = wrapPromptAsEvaluationScript(prompt)
       expect(validateEvaluationScript(script)).toBe(true)
     })
   })
