+++ packages/domain/evaluations/src/errors.ts
@@ -23,6 +23,16 @@ export class EvaluationManualRealignmentRateLimitedError extends Data.TaggedErro
   readonly httpMessage = "Manual evaluation realignment is temporarily rate limited"
 }
 
+export class EvaluationExecutionError extends Data.TaggedError("EvaluationExecutionError")<{
+  readonly message: string
+  readonly cause?: unknown
+}> {
+  readonly httpStatus = 422
+  get httpMessage() {
+    return this.message
+  }
+}
+
 export class LiveEvaluationExecutionError extends Data.TaggedError("LiveEvaluationExecutionError")<{
   readonly evaluationId: string
   readonly message: string
