+++ packages/domain/evaluations/src/index.ts
@@ -1,15 +1,11 @@
 export {
-  type EvaluationScriptExecution,
-  estimateEvaluationScriptCostMicrocents,
   evaluateDraftAgainstExamplesUseCase,
-  executeEvaluationScript,
-} from "./alignment-execution.ts"
-export { evaluateOptimizationCandidate } from "./alignment-optimization.ts"
+} from "./use-cases/alignment/evaluate-draft-against-examples.ts"
+export { evaluateOptimizationCandidate } from "./use-cases/optimization/evaluate-optimization-candidate.ts"
 export type {
   BaselineEvaluationExampleResult,
   BaselineEvaluationResult,
   CollectedEvaluationAlignmentExamples,
-  EvaluationAlignmentConversationMessage,
   GeneratedEvaluationDetails,
   GeneratedEvaluationDraft,
   HydratedEvaluationAlignmentExample,
@@ -17,28 +13,39 @@ export type {
   LoadedEvaluationAlignmentState,
   PersistEvaluationAlignmentResult,
   WriteEvaluationAlignmentJobStatusInput,
-} from "./alignment-types.ts"
+} from "./alignment/types.ts"
+export { collectAlignmentExamplesUseCase } from "./use-cases/alignment/collect-alignment-examples.ts"
+export { evaluateBaselineDraftUseCase } from "./use-cases/alignment/evaluate-baseline-draft.ts"
+export { evaluateIncrementalDraftUseCase } from "./use-cases/alignment/evaluate-incremental-draft.ts"
+export { generateBaselineDraftUseCase } from "./use-cases/alignment/generate-baseline-draft.ts"
+export { loadAlignmentStateUseCase } from "./use-cases/alignment/load-alignment-state.ts"
+export { persistAlignmentResultUseCase } from "./use-cases/alignment/persist-alignment-result.ts"
 export {
-  collectAlignmentExamplesUseCase,
-  evaluateBaselineDraftUseCase,
-  evaluateIncrementalDraftUseCase,
-  generateBaselineDraftUseCase,
-  loadAlignmentStateUseCase,
-  persistAlignmentResultUseCase,
-} from "./alignment-use-cases.ts"
-export {
-  CONVERSATION_PLACEHOLDER,
-  type ConversationMessage,
+  EVALUATION_CONVERSATION_PLACEHOLDER,
   EVALUATION_SCRIPT_RUNTIME_MODEL,
   EVALUATION_SCRIPT_RUNTIME_SYSTEM_PROMPT,
+  type EvaluationConversationMessage,
+  type EvaluationExecutionResult,
+  type EvaluationExecutionResultPayload,
+  type EvaluationIssueContext,
+  type EvaluationScriptExecution,
   type EvaluationScriptSchema,
+  type ExecuteEvaluationScriptWithAIError,
+  evaluationExecutionResultPayloadSchema,
+  evaluationExecutionResultSchema,
+  evaluationIssueContextSchema,
   evaluationRuntimeZod,
-  extractPromptFromScript,
-  formatConversationForPrompt,
-  generateBaselinePromptText,
+  estimateEvaluationScriptCostMicrocents,
+  executeEvaluationScript,
+  executeEvaluationScriptWithAI,
+  extractPromptFromEvaluationScript,
+  formatEvaluationConversationForPrompt,
+  toEvaluationConversationMessages,
+  toEvaluationExecutionResult,
   validateEvaluationScript,
-  wrapPromptAsScript,
-} from "./baseline-generation.ts"
+  wrapPromptAsEvaluationScript,
+} from "./runtime/evaluation-execution.ts"
+export { generateBaselinePromptText } from "./alignment/baseline-prompt.ts"
 export {
   ALIGNMENT_CURATED_DATASET_MAX_ROWS,
   ALIGNMENT_CURATED_DATASET_MIN_ROWS,
@@ -81,6 +88,7 @@ export {
   isPausedEvaluation,
 } from "./entities/evaluation.ts"
 export {
+  EvaluationExecutionError,
   EvaluationDeletedError,
   EvaluationManualRealignmentRateLimitedError,
   EvaluationNotFoundError,
@@ -116,7 +124,6 @@ export {
   parseStoredEvaluationAlignmentJobStatus,
   shouldSampleLiveEvaluation,
   softDeleteEvaluation,
-  toAlignmentConversationMessages,
   toLiveEvaluationDebounceMs,
   totalConfusionMatrixObservations,
   truncateEvaluationName,
@@ -135,7 +142,7 @@ export {
   liveEvaluationExecutionResultSchema,
   liveEvaluationIssueContextSchema,
   liveEvaluationResultPayloadSchema,
-} from "./live-execution.ts"
+} from "./use-cases/live/execute-live-evaluation.ts"
 export {
   DEFAULT_ALIGNMENT_EXAMPLE_LIMIT,
   type EvaluationAlignmentExample,
