+++ packages/domain/evaluations/src/use-cases/alignment/evaluate-baseline-draft.ts
@@ -1,0 +1,17 @@
+import type { HydratedEvaluationAlignmentExample } from "../../alignment/types.ts"
+import { evaluateDraftAgainstExamplesUseCase } from "./evaluate-draft-against-examples.ts"
+
+export const evaluateBaselineDraftUseCase = (input: {
+  readonly issueName: string
+  readonly issueDescription: string
+  readonly script: string
+  readonly positiveExamples: readonly HydratedEvaluationAlignmentExample[]
+  readonly negativeExamples: readonly HydratedEvaluationAlignmentExample[]
+}) =>
+  evaluateDraftAgainstExamplesUseCase({
+    issueName: input.issueName,
+    issueDescription: input.issueDescription,
+    script: input.script,
+    positiveExamples: input.positiveExamples,
+    negativeExamples: input.negativeExamples,
+  })
