+++ packages/domain/evaluations/src/use-cases/alignment/generate-baseline-draft.ts
@@ -1,0 +1,23 @@
+import { hashOptimizationCandidateText } from "@domain/optimizations"
+import { Effect } from "effect"
+import type { GeneratedEvaluationDraft } from "../../alignment/types.ts"
+import { generateBaselinePromptText } from "../../alignment/baseline-prompt.ts"
+import { defaultEvaluationTrigger } from "../../entities/evaluation.ts"
+import { wrapPromptAsEvaluationScript } from "../../runtime/evaluation-execution.ts"
+
+// TODO(eval-sandbox): restore LLM-based baseline generation for arbitrary scripts when sandbox
+// is available.
+export const generateBaselineDraftUseCase = (input: {
+  readonly issueName: string
+  readonly issueDescription: string
+}) =>
+  Effect.gen(function* () {
+    const promptText = generateBaselinePromptText(input.issueName, input.issueDescription)
+    const script = wrapPromptAsEvaluationScript(promptText)
+
+    return {
+      script,
+      evaluationHash: yield* Effect.tryPromise(() => hashOptimizationCandidateText(script)),
+      trigger: defaultEvaluationTrigger(),
+    } satisfies GeneratedEvaluationDraft
+  })
