+++ packages/domain/evaluations/src/use-cases/optimization/evaluate-optimization-candidate.ts
@@ -1,13 +1,7 @@
-import { AI } from "@domain/ai"
 import type { OptimizationCandidate, OptimizationTrajectory } from "@domain/optimizations"
 import { Effect } from "effect"
-import { executeEvaluationScript } from "./alignment-execution.ts"
-import type { HydratedEvaluationAlignmentExample } from "./alignment-types.ts"
-import {
-  EVALUATION_SCRIPT_RUNTIME_MODEL,
-  EVALUATION_SCRIPT_RUNTIME_SYSTEM_PROMPT,
-  type EvaluationScriptSchema,
-} from "./baseline-generation.ts"
+import type { HydratedEvaluationAlignmentExample } from "../../alignment/types.ts"
+import { executeEvaluationScriptWithAI } from "../../runtime/evaluation-execution.ts"
 
 // TODO(eval-sandbox): when sandbox is available, executeEvaluationScript will run arbitrary JS
 // and this function's structure will remain the same — it just calls executeEvaluationScript.
@@ -18,30 +12,14 @@ export const evaluateOptimizationCandidate = (input: {
   readonly issueDescription: string
 }) =>
   Effect.gen(function* () {
-    const ai = yield* AI
-    const services = yield* Effect.services<never>()
-    const execution = yield* Effect.tryPromise(() =>
-      executeEvaluationScript({
-        script: input.candidate.text,
-        conversation: input.example.conversation,
-        issue: {
-          name: input.issueName,
-          description: input.issueDescription,
-        },
-        generateStructuredObject: <T>(llmInput: {
-          readonly prompt: string
-          readonly schema: EvaluationScriptSchema<T>
-        }) =>
-          Effect.runPromiseWith(services)(
-            ai.generate({
-              ...EVALUATION_SCRIPT_RUNTIME_MODEL,
-              system: EVALUATION_SCRIPT_RUNTIME_SYSTEM_PROMPT,
-              prompt: llmInput.prompt,
-              schema: llmInput.schema,
-            }),
-          ),
-      }),
-    )
+    const execution = yield* executeEvaluationScriptWithAI({
+      script: input.candidate.text,
+      conversation: input.example.conversation,
+      issue: {
+        name: input.issueName,
+        description: input.issueDescription,
+      },
+    })
 
     const expectedPositive = input.example.label === "positive"
     const predictedPositive = execution.result.passed === false
